requirements.txt
streamlit
sqlmodel
sqlalchemy
psycopg[binary]
python-dotenv
pandas
plotly
